Refactor temperature and force into a separate modules
