+++ src/conversion/force/_force.js
@@ -1,0 +1,14 @@
+const shared = require('../shared');
+
+function forceMap(imperialInputs, metricTransform) {
+  const newtons = imperialInputs.map(metricTransform);
+
+  return shared.createMap(newtons, newtons => newtons, " N");
+}
+
+const metricForceUnits = [/newtons?/, /dynes?/];
+
+module.exports = {
+  "toMap" : forceMap,
+  "metricUnits" : metricForceUnits
+}
